requirements.txt
playwright>=1.53.0
python-dotenv>=1.1.1
cloudinary>=1.44.1
supabase>=2.17.0
openai>=1.95.1
requests>=2.32.4
yt-dlp>=2023.12.30
ffmpeg-python>=0.2.0
Pillow>=10.0.0 